requirements.txt
flask
pymongo
python-dotenv
pytest
black
flake8
pyswagger
pydantic
flasgger


